requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.4.32
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pytesseract==0.3.13
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
zstandard==0.25.0
